move tests so that they live in the same directory as the component
