+++ packages/tc-ui/src/primitives/relationship/e2e-create-relationships.cyp.js
@@ -1,4 +1,4 @@
-const { code } = require('../fixtures/mainTypeData.json');
+const { code } = require('../../../../../cypress/fixtures/mainTypeData.json');
 const {
 	populateMinimumViableFields,
 	populateParentTypeFields,
@@ -9,12 +9,8 @@ const {
 	visitEditPage,
 	visitMainTypePage,
 	save,
-} = require('../test-helpers');
-const { dropFixtures } = require('../../test-helpers/test-fixtures');
-
-const resetDb = async () => {
-	await dropFixtures(code);
-};
+	resetDb,
+} = require('../../test-helpers');
 
 describe('End-to-end - relationship creation', () => {
 	beforeEach(() => {
